[WIP] Preparations before card insertion
This is quite a big commit that I wrote in a code fever, where it
wouldn't really make sense to commit anything, but this state is
advanced enough and the next step big enough, that it makes sense to do
at least one.

In this commit stuff like communication with the WaniKani servers and
with Anki work flawlessly. The API has been written for Anki so that
decks can be created from scratch with models, templates, ... that is
otherwise required.

The templates are automatically filled from the resource folder and an
init can even update templates, styling and such.

It's a really powerful base which is necessary for the next step, which
is actually parsing the retrieved subjects and inserting them into the
deck as cards.

Credits for template code goes to:
https://github.com/FredericaBernkastel/WaniKani/tree/master

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>