working sync of kanji and radicals to Anki
With this all kanji and radical subjects are being synced successfully
to Anki, including images instead of characters for e.g. radicals and
cross-referencing of cards (e.g. radicals that make up a kanji)

There have been a few changes structurally. First and foremost subjects
are now in a subfolder and are classes that contain all their respective
functions. With this I can define an interface, allowing me to
standardize their interactions, allowing for more generalization in e.g.
wanideck.

Also the deck class now contain the bulk of functionality that is
required to interact with the Anki API in a manner I need, reducing the
complexity of wanideck.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>