[feat] assignments: only retrieve new assignments
This adds a new field to the metadata note and renames another one.
Let's hope nobody has used this program so far except for me, as this
requires manual renaming in anki if one doesn't want to have breaking
changes.

This gist of this change is that only new assignments are being
retrieved and added to anki. As this process currently does a forcibly
overwrite, I noticed when doing the `init` command I reset some of my
anki progress. So this change is to fight against that.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>