[feat] templates: remove spoilers & cleanup
- remove spoilers
Cards like vocab and kanji have a reading and a recognition task. The
back of both types of cards always contain everything from that note. So
helps learning and lookup if one really wants to know. But this also
meant that one would be spoiled with the answer of the respective other
card type. So now added spoiler elements where appropriate to hide the
information and only autoplay the audio on reading cards.

- cleanup of main.css
There still were a lot of style definitions that are not used anymore in
it.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>